implement getRobotModelString
It auto-detects whether a string is a file path or a XML string.
It normalizes the former into the latter and returns the string
and format (auto-detecting the format if requested, when the file
has a known extension)

This is meant to be used at the interface level, to normalize
inputs into a string/format pair before being processed by e.g.
loading the KDL tree.